requirements.txt
pyTelegramBotAPI
Flask
psutil
python-dotenv